地铁管理抽屉组件完成

+++ src/auth-provider.ts
@@ -1,3 +1,5 @@
+import {message} from "antd";
+
 const apiUrl = process.env.REACT_APP_API_URL;
 
 const sessionStorageKey = "jwtToken";
@@ -20,6 +22,7 @@ export const login = (data: { loginName: string; password: string }) => {
     body: JSON.stringify(data),
   }).then(async (response) => {
     if (response.ok) {
+      message.success("登陆成功")
       return handleUserResponse(await response.json());
     } else {
       return Promise.reject(await response.json());
@@ -32,8 +35,10 @@ export const logout = async () => {
     method: "POST", headers: {
       "Authorization": getToken() ? `${getToken()}` : ''
     }
-  }).then(() => {
-    window.sessionStorage.clear()
-    window.location.href = window.location.origin
+  }).then((res) => {
+    if (res.ok) {
+      window.sessionStorage.clear()
+      window.location.href = window.location.origin
+    }
   })
 }
